Separated the UI Object into its own Module.
The UI Object relies on the Layers Object, so a new module was created for Layers.
The UI Object also relies on the Property Object, so a module was created for Property as well.

Because the Tool Object relied on the UI Object, the Tool Object was now able to be completed. The reverse is also true (UI relies on Tool Object).
