+++ core.js
@@ -1,302 +1,15 @@
 import { cache, pressed } from './Cache.js';
 import { doc } from './SetUp.js';
-
-import { colors } from './Tab/Color.js';
 import { tabStates, util } from './Tabs.js';
+import { ui } from './UI.js';
+
+import { tool } from './Tab/Tool.js'
+import { colors } from './Tab/Color.js';
+import { property } from './Tab/Property.js';
+import { layers } from './Tab/Layer.js';
 
 $(document).ready(function() {
 
-
-  var ui = {
-    cursor: function(type) {
-      $('#editor').css('cursor',type);
-      $('.selection').css('cursor',type);
-    },
-    resizeHandles: function(toggle) {
-      if (!toggle) {
-        $('.selection').addClass('selecting');
-      }
-      if (toggle) {
-        $('.selection').removeClass('selecting');
-      }
-      /*if (toggle == 'selection') {
-        $('selection *').css('display','block');
-      }*/
-    },
-    hideTools: function(toggle) {
-      if (toggle == true) {
-        $('.selection, .tools').addClass('hide');
-        ui.cursor('grab');
-      } else if (toggle == 'selection') {
-        $('.selection').addClass('hide');
-      } else {
-        $('.selection, .tools').removeClass('hide');
-        if (tool.type != 'selection') {
-          ui.cursor('crosshair');
-        }
-      }
-    },
-    scrubber: function(toggle,direction) {
-      if (toggle == false) {
-        if ($('.propertyScrubber').hasClass('show')) {
-          $('.propertyScrubber').removeClass('show');
-          layers.update();
-        }
-      } else if (!$('.propertyScrubber').hasClass('show')) {
-        if (direction == 'down') {
-          $('.propertyScrubber').removeClass('show').addClass('down');
-          $('.propertyScrubber').css({
-            'top': cache.btnArea.bottom,
-            'left': cache.btnArea.left + 2
-          });
-          setTimeout(function() {
-            $('.propertyScrubber').addClass('show').css({
-              'top': cache.btnArea.bottom
-            });
-          }, 1);
-        } else {
-          $('.propertyScrubber').removeClass('show').removeClass('down');
-          $('.propertyScrubber').css({
-            'top': cache.btnArea.top,
-            'left': cache.btnArea.left + 2
-          });
-          setTimeout(function() {
-            $('.propertyScrubber').addClass('show').css({
-              'top': cache.btnArea.top - 190
-            });
-          }, 1);
-        }
-        $('.propertyScrubber .scrub').css({
-          'background-color': property.scrubberTo.attr('stroke') ? tool.stroke : 'white'
-        });
-      }
-    },
-    location: function(btnName) {
-      var btn;
-      if (typeof btnName == 'string') {
-        btn = {
-          top: $('[aria-label="'+btnName+'"]').offset().top,
-          bottom: $('[aria-label="'+btnName+'"]').offset().top + $('[aria-label="'+btnName+'"]').height(),
-          left: $('[aria-label="'+btnName+'"]').offset().left,
-          right: $('[aria-label="'+btnName+'"]').offset().left + $('[aria-label="'+btnName+'"]').width()
-        }
-      } else {
-        btn = {
-          top: btnName.offset().top,
-          bottom: btnName.offset().top + btnName.height(),
-          left: btnName.offset().left,
-          right: btnName.offset().left + btnName.width()
-        }
-      }
-      return (cache.stop[0] > btn.left && cache.stop[0] < btn.right && cache.stop[1] > btn.top && cache.stop[1] < btn.bottom);
-    },
-    cursorFeedback: function() { // Depending on where the user is dragging the color, it will change styles if it over a droppable area
-      if (tabStates.index[tabStates.index.length-1] != 'properties' && this.location($('.color'))) {
-        if (this.location('saved') || this.location('tuner')) {
-          $('.colorDragging').addClass('drop-shrink');
-          this.location('saved') ? $('[aria-label="saved"]').addClass('validDrop') : $('[aria-label="tuner"]').addClass('validDrop');
-        } else {
-          $('.colorDragging').removeClass('drop-shrink');
-          $('.category .material-icons').removeClass('validDrop');
-        }
-        if ($('[aria-label="tuner"]').hasClass('select') || $('[aria-label="saved"]').hasClass('select')) {
-          if (this.location('palette')) {
-            $('.colorDragging').addClass('remove');
-          } else {
-            $('.colorDragging').removeClass('remove');
-          }
-        }
-      } else if (this.location($('.properties'))) {
-        tabStates.indexUp($('.properties'));
-        if ((this.location('stroke') || this.location('fill'))) {
-          $('.colorDragging').addClass('drop-grow');
-          $('.colorDragging').attr('data-overicon','border_color');
-          if (this.location('fill')) {
-            $('.colorDragging').attr('data-overicon','format_color_fill');
-          }
-        } else {
-          $('.colorDragging').removeClass('drop-grow');
-        }
-      } else if (this.location($('.color'))) {
-        $('.colorDragging').removeClass('drop-grow');
-        tabStates.indexUp($('.color'));
-      } else {
-        $('.colorDragging').removeClass('drop-grow')
-      }
-    },
-    sendColor: function() { //
-      if (tabStates.index[tabStates.index.length-1] != 'properties' && this.location($('.color'))) {
-        if (this.location('saved')) {
-          colors.push('saved',cache.dragColor);
-          $('.colorDragging').addClass('shrink-out');
-          var savedBtn = $('[aria-label="saved"]');
-          $('.colorDragging').css({
-            'left': savedBtn.offset().left + savedBtn.width()/2 - $('.colorDragging').width()/2,
-            'top': savedBtn.offset().top + savedBtn.height()/2 - $('.colorDragging').height()/2
-          });
-        } else if (this.location('tuner')) {
-          colors.push('tuner',cache.dragColor);
-          $('.colorDragging').addClass('shrink-out');
-          var tunerBtn = $('[aria-label="tuner"]');
-          $('.colorDragging').css({
-            'left': tunerBtn.offset().left + tunerBtn.width()/2 - $('.colorDragging').width()/2,
-            'top': tunerBtn.offset().top + tunerBtn.height()/2 - $('.colorDragging').height()/2
-          });
-        } else if (this.location('palette') && ($('[aria-label="tuner"]').hasClass('select') || $('[aria-label="saved"]').hasClass('select'))) {
-          if ($('[aria-label="tuner"]').hasClass('select')) {
-
-          } else if ($('[aria-label="saved"]').hasClass('select')) {
-
-          }
-          $('.colorDragging').css({
-            'opacity': 0,
-            'filter': 'blur(5px)'
-          });
-        } else {
-          $('.colorDragging').css({
-            'left': cache.start[0],
-            'top': cache.start[1]
-          }).removeClass('dragging').addClass('scrolling');
-        }
-      } else if (this.location($('.properties'))) {
-        var strokeBtn = $('[aria-label="stroke"]');
-        if (this.location('stroke') || this.location('fill')) {
-          if (this.location('stroke')) {
-            tool.stroke = cache.dragColor;
-            cache.ele.attr('stroke',cache.dragColor);
-            if (cache.ele.css('stroke')) {
-              cache.ele.css('stroke','');
-            }
-            $('[aria-label="stroke"]').css('color',cache.dragColor);
-          }
-          if (this.location('fill')) {
-            strokeBtn = $('[aria-label="fill"]');
-            tool.fill = cache.dragColor;
-            cache.ele.attr('fill',cache.dragColor);
-            if (cache.ele.css('fill')) {
-              cache.ele.css('fill','');
-            }
-            $('[aria-label="fill"]').css('color',cache.dragColor);
-          }
-          colors.push('recent',cache.dragColor);
-          property.colorTo = strokeBtn;
-          $('.colorDragging').css({
-            'left': strokeBtn.offset().left + strokeBtn.width()/2 - $('.colorDragging').width()/2,
-            'top': strokeBtn.offset().top + strokeBtn.height()/2 - $('.colorDragging').height()/2
-          }).addClass('grow-out');
-
-        } else {
-          tabStates.indexUp($('.color'));
-          $('.colorDragging').css({
-            'left': cache.start[0],
-            'top': cache.start[1]
-          }).removeClass('dragging').addClass('scrolling');
-        }
-        /*tool[property.colorTo] = $('.colorDragging').css('background-color');
-        cache.ele.attr(property.colorTo,$('.colorDragging').css('background-color'));*/
-      } else {
-        $('.colorDragging').css({
-          'left': cache.start[0],
-          'top': cache.start[1]
-        }).removeClass('dragging').addClass('scrolling');
-      }
-      setTimeout(function() {
-        $('.colorDragging').remove();
-        $('.category .material-icons').removeClass('validDrop');
-      }, 500);
-    },
-    showDropArea: function(layer) {
-      if (cache.stop[1] < layer.offset().top + 12) {
-        $('.layers div').removeClass('drop-below drop-group');
-        if (!($('.layers .selected').length >= 1 && layers.current.hasClass('selected') && layer.hasClass('selected'))) {
-          layer.addClass('drop-above');
-        }
-      } else if (cache.stop[1] > layer.offset().top + layer.outerHeight() - 12) {
-        $('.layers div').removeClass('drop-above drop-group');
-        if (!($('.layers .selected').length >= 1 && layers.current.hasClass('selected') && layer.hasClass('selected'))) {
-          layer.addClass('drop-below');
-        }
-      } else if (cache.stop[1] >= layer.offset().top + 12 && cache.stop[1] <= layer.offset().top + layer.outerHeight() - 12) {
-        $('.layers div').removeClass('drop-above drop-below');
-        layer.addClass('drop-group');
-      } else {
-        $('.layers div').removeClass('drop-above drop-below drop-group');
-      }
-      if (!$('.draggingLayer').length) {
-        layers.current.clone().appendTo('.tools').addClass('draggingLayer');
-      } else {
-        $('.draggingLayer').css({
-          'left': cache.stop[0] - layer.outerWidth()/2
-        });
-
-        if ($('.draggingLayer').hasClass('layerAboveCursor')) {
-          $('.draggingLayer').css({
-            'top': cache.stop[1] - layer.outerHeight()
-          });
-        } else {
-          $('.draggingLayer').css({
-            'top': cache.stop[1]
-          })
-        }
-      }
-    },
-    scrollbar: function(show) {
-      if (show) {
-        $('.scrollbar').addClass('show')
-      } else {
-        $('.scrollbar').removeClass('show')
-      }
-    },
-    colorListPreview: function(scheme) {
-      if (scheme) {
-        var colorsPreview = $('.schemes ul li.' + scheme + ' .colorsPreview');
-        colorsPreview.html('');
-
-      } else {
-        $('.schemes ul li').prepend('<div class="colorsPreview"></div><div class="colorsPreviewScroll"></div>');
-        $('.schemes ul li').each(function() {
-          var scheme = $(this).attr('data-scheme');
-          var array = colors.collections[scheme].colors;
-          var empty = colors.collections[scheme].empty;
-          var skipOver = Math.floor((array.length - empty) / 49);
-          skipOver == 0 ? skipOver = 1 : null;
-          for (var i = 0; i < array.length; i++) {
-            var color = array[i].split(' ')[0];
-            if (color.trim() != '') {
-              if (i % skipOver == 0) {
-                $(this).children('.colorsPreview').append('<div style="background-image: linear-gradient(to bottom, '+color+' 16px, transparent 48px)"></div>');
-              }
-              $(this).children('.colorsPreviewScroll').append('<div style="background-image: linear-gradient(to bottom, '+color+' 16px, transparent 48px)"></div>');
-            }
-          }
-        });
-      }
-    },
-    infocolor: function(action) {
-      $('infocolor').removeClass('quickhide');
-      if (action == true) {
-        $('.infoPanel').addClass('show');
-      } else if (action == false) {
-        $('.infoPanel').removeClass('show');
-      } else {
-        $('.infocolor').addClass('quickHide');
-        $('.infoPanel').removeClass('show');
-      }
-    },
-    selectScheme: function(ele) {
-      if ($('.schemes').hasClass('view')) {
-        //ele.addClass('transform-top');
-        colors.scheme = ele.attr('data-scheme');
-        $('.color input.user').val($(this).text());
-        $('.schemes li').removeClass('selected');
-        $('.schemes [data-scheme="'+colors.scheme+'"]').addClass('selected');
-        $('.swatches').attr('data-scheme',colors.scheme);
-        colors.scrolled.palette = 0;
-        colors.draw();
-        $('.schemes').removeClass('view');
-      }
-    }
-  }
 
   $(document).contextmenu(function(e) {
     e.preventDefault();
@@ -386,74 +99,8 @@ $(document).ready(function() {
     }
   });
   
-  $('#autosize').change(function() {
-    if ($(this).prop('checked')) {
-      $('.size input').attr('disabled', 'disabled');
-      $('.size input:eq(0)').val($(document).width());
-      $('.size input:eq(1)').val($(document).height());
-    } else {
-      $('.size input').removeAttr('disabled');
-    }
-    $('.create').addClass('show');
-  });
 
 
-
-
-
-
-  var tool = {
-    toolType: 'line',
-    name: false,
-    prevName: false,
-    set type(kind) {
-      $('.shapes div').removeClass('selected');
-      ui.hideTools(false);
-      if (kind == this.toolType) {
-        this.toolType = 'selection';
-        ui.cursor('default');
-      } else if (kind != 'selection') {
-        this.toolType = kind;
-        ui.cursor('crosshair');
-        $('#'+kind).addClass('selected');
-      } else {
-        this.toolType = 'selection';
-        ui.cursor('default');
-      }
-      if (kind == 'image') {
-        if (tool.imageIndex == -1) {
-          $('#inputFile').trigger('click');
-          tool.imageIndex = tool.prevIndexIMG;
-        }
-        this.toolType = 'selection';
-        ui.hideTools('selection');
-        ui.cursor('copy');
-      }
-      if (kind == 'drag' && !pressed.spaceBar) {
-        this.toolType = 'selection';
-        ui.hideTools(true);
-        ui.cursor('drag');
-      }
-      if (kind == 'animate') {
-        this.toolType = 'selection';
-      }
-      this.name = kind;
-    },
-    get type() {
-      return this.toolType;
-    },
-    stroke: 'black',
-    strokeWidth: 10,
-    strokeOpacity: 1,
-    fillOpacity: 1,
-    paintOrder: 'fill',
-    fill: 'black',
-    presets: [],
-    strokeDashOffset: 0,
-    strokeDashArray: 0,
-    images: [],
-    imageIndex: -1
-  }
 
 
   var translateX = 0;
@@ -1683,132 +1330,8 @@ $(document).ready(function() {
       }
     }
 
-    var preview; // including the variable outside the layers object seems to help in preventing glitches when dragging layers
-    var layers = {
-      all: [],
-      update: function() {
-        $('.layers .all').html('');
-        colors.picker = [];
-        for (var i = $('#editor').children().length-1; i >= 0; i--) { // Thi = {s loops through all of the SVG layers and renders them in the layers tab
-          var data = {
-            html: $('#editor').children().eq(i)[0].outerHTML,
-            type: $('#editor').children().eq(i)[0].outerHTML.replace('<','').split(' ')[0],
-            id: $('#editor').children().eq(i).attr('id').length ? $('#editor').children().eq(i).attr('id') : null
-          }
-          var fillColor, strokeColor, stroke, target;
-          if (data.type == 'circle' || data.type == 'ellipse' || data.type == 'rect' ||
-              data.type == 'line' || data.type == 'path' || data.type == 'polygon' || data.type == 'g') {
-            var ele = $('#' + data.id);
-            var stroke = ele.attr('stroke-width');
-            var width = ele[0].getBBox().width + stroke/2*2;
-            var height = ele[0].getBBox().height + stroke/2*2;
-            var transX = -ele[0].getBBox().x + stroke/2;
-            var transY = -ele[0].getBBox().y + stroke/2;
-            $('.layers .all').append('<div id="'+data.id+'">'+
-            '<svg id="'+data.id+'" viewBox="0 0 '+width+' '+height+'" height="50px" width="50px" preserveAspectRatio="xMidYMid meet"><g>' +
-            data.html +
-            '</g></svg>' +
-            '</div>');
-            preview = '.layers .all div[id="'+data.id+'"] g';
-            var parent = '.layers .all div[id="'+data.id+'"]';
-            stroke = parseFloat($(target).attr('stroke-width'));
-            if ($(preview).attr('fill')) {
-              fillColor = $(preview).attr('fill').toUpperCase();
-            } else {
-              fillColor = $(preview).css('fill').toUpperCase();
-            }
-            if ($(preview).attr('stroke')) {
-              strokeColor = $(preview).attr('stroke').toUpperCase();
-            } else {
-              strokeColor = $(preview).css('stroke').toUpperCase();
-            }
-            if ($(preview).attr('display') == 'none') {
-              $(parent).addClass('hidden');
-            }
-            $(preview).attr('display','block');
-            colors.push('picker',strokeColor);
-            colors.push('picker',fillColor);
-            $(preview).attr({
-              'transform': 'translate('+transX+','+transY+')'
-            });
-            $(preview).css({
-              'transform': '',
-              'matrix': ''
-            });
-          }
-        }
-      },
-      moveUp: function(top) {
-        if (top) {
-          cache.ele.detach().appendTo('#editor');
-        } else {
-          var index = cache.ele.index();
-          var nextEle = cache.ele.next();
-          if (nextEle.length) {
-            cache.ele.detach().insertAfter(nextEle);
-          }
-        }
-        this.update();
-      },
-      moveBack: function(bottom) {
-        if (bottom) {
-          cache.ele.detach().prependTo('#editor');
-        } else {
-          var index = cache.ele.index();
-          var prevEle = cache.ele.prev();
-          if (prevEle.length) {
-            cache.ele.detach().insertBefore(prevEle);
-          }
-        }
-        this.update();
-      },
-      drop: function(layer) {
-        if (layers.current.hasClass('selected') && $('.layers .selected').length > 1) {
-          if (!layer.hasClass('selected')) {
-            if (layer.hasClass('drop-above')) {
-              $('.layers .selected').insertBefore(layer);
-              $('.layers .selected').each(function() {
-                $('#editor #' + $(this).attr('id')).detach().insertAfter($('#editor #' + layer.attr('id')));
-              });
-            } else if (layer.hasClass('drop-below')) {
-              $('.layers .selected').insertAfter(layer);
-              $('.layers .selected').each(function() {
-                $('#editor #' + $(this).attr('id')).detach().insertBefore($('#editor #' + layer.attr('id')));
-              });
-            } else if (layer.hasClass('drop-group')) {
 
-            }
-          }
 
-        } else if (layers.current.attr('id') != layer.attr('id')) {
-          if (layer.hasClass('drop-above')) {
-            layers.current.detach().insertBefore(layer);
-            $('#editor #' + layers.current.attr('id')).detach().insertAfter($('#editor #' + layer.attr('id')));
-          } else if (layer.hasClass('drop-below')) {
-            layers.current.detach().insertAfter(layer); // insertAfter on this line and below are different because layers are show bottom to top in layers preview
-            $('#editor #' + layers.current.attr('id')).detach().insertBefore($('#editor #' + layer.attr('id')));
-          } else if (layer.hasClass('drop-group')) {
-
-          }
-        }
-        console.log(layer);
-        console.log(layers.current);
-        //if ()
-
-        /*if (layer.hasClass('drop-above')) {
-          $('.layers .selected').each(function() {
-            $('#editor ' + $(this).attr('id')).detach().index(layer.index());
-          });
-        } else if (layer.hasClass('drop-below')) {
-          $('.layers .selected').each(function() {
-            $('#editor ' + $(this).attr('id')).detach().index(layer.index() + 1);
-          });
-        } else {
-
-        }*/
-
-      }
-    }
 
     $('.layers .all').on('mouseenter', 'div', function(e) {
       /*cache.ele = $(e.target).attr('id');
@@ -1900,78 +1423,7 @@ $(document).ready(function() {
       }
     });
 
-    var property = {
-      dynamicUI: function() {
 
-      },
-      colorTo: 'stroke', // What the color applies to
-      showOptions: function() {
-
-      },
-      setNumValue: function() { // Change the value of a property
-        var property = this.scrubberTo;
-        var valueChange = cache.stop[1] - cache.start[1];
-        var scrubAmount = 0;
-        if (!$('.propertyScrubber').hasClass('down')) {
-          valueChange = cache.start[1] - cache.stop[1];
-        }
-        if (cache.ele.attr('stroke-opacity') == '0') {
-          cache.ele.attr('stroke-opacity',1);
-        }
-        if (cache.ele) {
-          switch(property.attr('aria-label')) {
-            case 'stroke':
-              if (!$('.propertyScrubber').hasClass('show')) {
-                if (cache.ele.attr('stroke-width')) {
-                  this.value = Math.round(parseFloat(cache.ele.attr('stroke-width')));
-                } else {
-                  this.value = 0;
-                }
-              }
-              valueChange = this.value + valueChange;
-              var value;
-              if (valueChange < 100) {
-                value = valueChange/10;
-                scrubAmount = value * 10;
-              } else {
-                value = Math.round((valueChange)-90);
-                scrubAmount = value + 90;
-              }
-              if (value > 100) {
-                value = 100;
-                scrubAmount = 190;
-                cache.start = [cache.stop[0],cache.stop[1]];
-                this.value = 100;
-              }
-              if (value < 0) {
-                value = 0;
-                scrubAmount = 0;
-                cache.start = [cache.stop[0],cache.stop[1]];
-                this.value = 0;
-              }
-              cache.ele.attr('stroke-width',value);
-              if (value < 10) {
-                if (value != 0) {
-                  value = value.toFixed(1);
-                }
-              }
-              $('.propertyScrubbwer').attr({
-                'data-value': value
-              }).removeClass('smallText');
-              $('.scrub').css({
-                'height': scrubAmount
-              });
-              tool.strokeWidth = value;
-              break;
-
-          }
-        } else {
-          $('.propertyScrubber').attr({
-            'data-value': 'stroke width'
-          }).addClass('smallText');
-        }
-      }
-    }
 
 
     $('.properties div').click(function() {
